+++ eyes.sdk.core/src/main/java/com/applitools/eyes/exceptions/NewTestException.java
@@ -8,7 +8,7 @@ import com.applitools.eyes.TestResults;
 public class NewTestException extends TestFailedException {
 
     public NewTestException(TestResults testResults, String scenarioIdOrName, String appIdOrName){
-        super(String.format("'%s' of '%s'. Please approve the new baseline at %s",
+        super(testResults, String.format("'%s' of '%s'. Please approve the new baseline at %s",
                 scenarioIdOrName,
                 appIdOrName,
                 testResults.getUrl()));